+++ park/sec1/INF0103.java
@@ -1,4 +1,4 @@
-package park;
+package park.sec1;
 
 import java.util.Scanner;
