+++ park/sec1/INF0106T1.java
@@ -1,4 +1,4 @@
-package park;
+package park.sec1;
 
 import java.util.Scanner;
