+++ park/sec1/INF0107.java
@@ -1,31 +1,30 @@
-package park;
+package park.sec1;
 
 import java.util.Scanner;
 
-public class INF0108 {
+public class INF0107 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        String s = sc.nextLine().toUpperCase().replaceAll("[^A-Z]"," ");
-        s = s.replace(" ","");
+        String s = sc.next().toUpperCase();
 
-        INF0108 inf = new INF0108();
+        INF0107 inf = new INF0107();
         System.out.println(inf.solution(s));
     }
 
     public String solution(String s) {
         String answer = "YES";
+        int lt = 0, rt = s.length() - 1;
         char[] arr = s.toCharArray();
 
-        int lt = 0, rt = s.length() - 1;
         while(lt < rt) {
-            if(arr[lt] != arr[rt]){
+            if(arr[lt] == arr[rt]) {
+                lt++;
+                rt--;
+            }else {
                 answer = "NO";
                 break;
             }
-            lt++;
-            rt--;
         }
-
         return answer;
     }
 }
