+++ park/sec1/INF0108.java
@@ -1,30 +1,31 @@
-package park;
+package park.sec1;
 
 import java.util.Scanner;
 
-public class INF0107 {
+public class INF0108 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        String s = sc.next().toUpperCase();
+        String s = sc.nextLine().toUpperCase().replaceAll("[^A-Z]"," ");
+        s = s.replace(" ","");
 
-        INF0107 inf = new INF0107();
+        INF0108 inf = new INF0108();
         System.out.println(inf.solution(s));
     }
 
     public String solution(String s) {
         String answer = "YES";
-        int lt = 0, rt = s.length() - 1;
         char[] arr = s.toCharArray();
 
+        int lt = 0, rt = s.length() - 1;
         while(lt < rt) {
-            if(arr[lt] == arr[rt]) {
-                lt++;
-                rt--;
-            }else {
+            if(arr[lt] != arr[rt]){
                 answer = "NO";
                 break;
             }
+            lt++;
+            rt--;
         }
+
         return answer;
     }
 }
