+++ park/sec1/INF0112.java
@@ -1,4 +1,4 @@
-package park;
+package park.sec1;
 
 import java.util.Arrays;
 import java.util.Scanner;
